Phase 3 Task 2: Fix import statements and missing type exports
- Added ModificationType export to adaptation.ts module
- Moved RunnerAttribute type from methodology-recommendation-engine.ts to types.ts
- Fixed WORKOUT_TEMPLATES import in methodology-customization-engine.ts from constants to workouts
- Updated PlanModification interface to use ModificationType
- All import-related errors resolved
- Error count reduced from 499 to 496 (3 errors fixed)
